+++ packages/jest-cli/src/__tests__/watch.test.js
@@ -101,7 +101,29 @@ describe('Watch mode flows', () => {
   });
 
   it('Runs Jest once by default and shows usage', () => {
-    watch(globalConfig, contexts, pipe, hasteMapInstances, stdin);
+    jest.unmock('jest-util');
+    const util = require('jest-util');
+    util.isInteractive = true;
+
+    const ci_watch = require('../watch').default;
+    ci_watch(globalConfig, contexts, pipe, hasteMapInstances, stdin);
+    expect(runJestMock.mock.calls[0][0]).toMatchObject({
+      contexts,
+      globalConfig,
+      onComplete: expect.any(Function),
+      outputStream: pipe,
+      testWatcher: new TestWatcher({isWatchMode: true}),
+    });
+    expect(pipe.write.mock.calls.reverse()[0]).toMatchSnapshot();
+  });
+
+  it('Runs Jest in a non-interactive environment not showing usage', () => {
+    jest.unmock('jest-util');
+    const util = require('jest-util');
+    util.isInteractive = false;
+
+    const ci_watch = require('../watch').default;
+    ci_watch(globalConfig, contexts, pipe, hasteMapInstances, stdin);
     expect(runJestMock.mock.calls[0][0]).toMatchObject({
       contexts,
       globalConfig,
@@ -128,7 +150,12 @@ describe('Watch mode flows', () => {
   });
 
   it('shows prompts for WatchPlugins in alphabetical order', async () => {
-    watch(
+    jest.unmock('jest-util');
+    const util = require('jest-util');
+    util.isInteractive = true;
+
+    const ci_watch = require('../watch').default;
+    ci_watch(
       Object.assign({}, globalConfig, {
         rootDir: __dirname,
         watchPlugins: [watchPlugin2Path, watchPluginPath],