+++ packages/jest-cli/src/reporters/default_reporter.js
@@ -14,9 +14,8 @@ import type {GlobalConfig, Path, ProjectConfig} from 'types/Config';
 import type {Test} from 'types/TestRunner';
 import type {ReporterOnStartOptions} from 'types/Reporters';
 
-import {clearLine, getConsoleOutput} from 'jest-util';
+import {clearLine, getConsoleOutput, isInteractive} from 'jest-util';
 import chalk from 'chalk';
-import isCI from 'is-ci';
 import BaseReporter from './base_reporter';
 import Status from './Status';
 import getResultHeader from './get_result_header';
@@ -26,8 +25,6 @@ type write = (chunk: string, enc?: any, cb?: () => void) => boolean;
 type FlushBufferedOutput = () => void;
 
 const TITLE_BULLET = chalk.bold('\u25cf ');
-
-const isInteractive = process.stdin.isTTY && !isCI;
 
 export default class DefaultReporter extends BaseReporter {
   _clear: string; // ANSI clear sequence for the last printed status